Implemented forge energy capability.

+++ src/main/java/org/halvors/quantum/client/gui/machine/GuiChemicalExtractor.java
@@ -66,7 +66,7 @@ public class GuiChemicalExtractor extends GuiContainerBase {
         drawSlot(79, 49, SlotType.BATTERY);
         drawSlot(52, 24);
         drawSlot(106, 24);
-        drawBar(75, 24, tile.timer / TileChemicalExtractor.tickTime);
+        drawBar(75, 24, (float) tile.timer / (float) TileChemicalExtractor.tickTime);
         drawMeter(8, 18, tile.tank.getInputTank().getFluidAmount() / tile.tank.getInputTank().getCapacity(), tile.tank.getInputTank().getFluid());
         drawSlot(24, 18, SlotType.LIQUID);
         drawSlot(24, 49, SlotType.LIQUID);
